aws-server/oseduwd_vanilla_aws_server1/source/handlers/.requirements.txt
json
requests
base64
csv
boto3
botocore.exceptions